requirements.txt
openai
langchain
pinecone-client
pinecone-text
google-search-results
tiktoken
python-dotenv
aiohttp
